docs+packaging: audit follow-ups — version from pyproject, min macOS 11, source re-attribution, native-integration notes
- packaging/joyglide_macos.spec: CFBundleVersion/ShortVersionString read
  from pyproject.toml (Info.plist had drifted to 0.1.0 at project 0.1.2);
  LSMinimumSystemVersion 10.15 → 11.0 for the arm64-only binary.
- player.py: drop gc.collect() in __init__/disconnect (full GC pass
  stalls every other controller's pump); joycon.py: battery_current_ma
  is a float.
- main.py: docstrings — one shared bg loop (not one per player),
  start_notify raises ValueError in bleak ≥2, no side-specific subscribe.
- ARCHITECTURE.md / README / NOTICE / CLAUDE.md: TropicalCyclone repo is
  deleted → cite Nadeflore/switch2-controllers (same code); BlueRetro
  archived; add JoyCon2Mac + joycon2cpp 0x37 nuance; cmd 0x03 subcommand
  names per ndeadly 2026-04; rewrite the Native macOS Integration section
  for NSActivity flags, thread QoS, NSScreen refresh rate, mouse deltas,
  hotkey tap self-healing; Known Limitations #4 (shared loop).
- CHANGELOG: [Unreleased] section for the 2026-08-29 audit.

+++ joycon.py
@@ -71,7 +71,7 @@ class JoyCon:
         # second to avoid UI churn.
         self.battery_mv: int | None        = None
         self.battery_pct: int | None       = None
-        self.battery_current_ma: int | None = None
+        self.battery_current_ma: float | None = None
         self.battery_charging = False
         self.battery_full     = False
         self._battery_last_ts = 0.0
